requirements.txt
asgiref==3.3.1
Django==3.1.4
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
Pillow==8.0.1
Pygments==2.7.3
PyJWT==2.0.0
pytz==2020.5
sqlparse==0.4.1
